pyproject.toml
[tool.poetry]
name = "foundry-backup-automation"
version = "0.1.0"
description = ""
authors = ["Joel Júnior <joel.j@escolar.ifrn.edu.br>"]
readme = "README.md"
packages = [{include = "foundry_backup_automation"}]

[tool.poetry.dependencies]
python = "^3.10"
paramiko = "^3.2.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"


[tool.poetry.group.development.dependencies]
blue = "^0.9.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
